backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LLM APIs
anthropic==0.40.0
openai==1.58.1  # For OpenRouter support

# Data validation
pydantic==2.10.5
pydantic-settings==2.7.1

# Markdown and YAML parsing
python-frontmatter==1.1.0
pyyaml==6.0.2

# File operations
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
